+++ CHANGELOG.md
@@ -30,6 +30,7 @@
   * Support argument value `type=user` in the action `get_art`
   * Support argument `size` in the action `get_art`
   * Add property `user` to all playlist results
+  * Allow `limit` values greater than 5000 and value "none" for unlimited result sets. Default is still 5000.
 
 ### Fixed
 - Playback stopping when clicking the title "0 results" of an empty advanced search result